refactor(solace-message-client): Avoid using minified file.
To provide better stack traces when debugging.
