+++ projects/solace-message-client/src/lib/multi-solace-message-client.spec.ts
@@ -1,7 +1,7 @@
 import {SolaceMessageClient} from './solace-message-client';
 import {createEnvironmentInjector, EnvironmentInjector, inject, runInInjectionContext, Type} from '@angular/core';
 import {provideSolaceMessageClient} from './solace-message-client.provider';
-import {LogLevel, SessionEventCode} from 'solclientjs';
+import {LogLevel, SessionEventCode} from 'solclientjs/lib-browser/solclient-full';
 import {SessionFixture} from './testing/session.fixture';
 import {provideSession} from './testing/session-provider';
 import {TestBed} from '@angular/core/testing';
