+++ projects/solace-message-client/src/lib/solace-message-client-connect.spec.ts
@@ -1,6 +1,6 @@
 import {MessageEnvelope, SolaceMessageClient} from './solace-message-client';
 import {TestBed} from '@angular/core/testing';
-import {SessionEventCode} from 'solclientjs';
+import {SessionEventCode} from 'solclientjs/lib-browser/solclient-full';
 import {SessionFixture} from './testing/session.fixture';
 import {provideSession} from './testing/session-provider';
 import {provideSolaceMessageClient} from './solace-message-client.provider';
