+++ projects/solace-message-client/src/lib/solace-message-client-dispose.spec.ts
@@ -1,7 +1,7 @@
 import {SolaceMessageClient} from './solace-message-client';
 import {createEnvironmentInjector, EnvironmentInjector} from '@angular/core';
 import {provideSolaceMessageClient} from './solace-message-client.provider';
-import {SessionEventCode} from 'solclientjs';
+import {SessionEventCode} from 'solclientjs/lib-browser/solclient-full';
 import {SessionFixture} from './testing/session.fixture';
 import {provideSession} from './testing/session-provider';
 import {TestBed} from '@angular/core/testing';
