+++ projects/solace-message-client/src/solclientjs.d.ts
@@ -1,0 +1,4 @@
+declare module 'solclientjs/lib-browser/solclient-full' {
+  import solace = require('solclientjs/lib-browser/index');
+  export = solace;
+}
